Modified how DataViewTypes are registered (#4187)
* Modified how DataViewTypes are registered

The DataViewManager registers DataViewTypes to determine how to
represent the the data within ML.Net. However, there was a bug in how
types were being queried. Types that were being registered via an
Attribute were being added to the manager with only that Attribute.
However, when queried from the manager all custom Attributes were being
passed.
To solve this we need to only pass custom Attributes that are relevant
to the Manager. All Type attributes now inherit from a single
TypeAttribute class such that we can filter custom attributes for
relevant types.

* Updated based on review comments to remove TypeAttribute

Since the DataViewManager works only on DataViewTypeAttributes, we do
not need to subclass attributes, but only the filter on the
DataViewTypeAttribute instead.

* Added test case to match initial bug report

While the unit test added previously addressed the underlying issue,
this test case proves that the issue in Bug #4121 is now completely
resolved, and there should be no further issues.

* change to use existing model

Did not want to polute the repo with large models so was able to find an
existing model which had an appropriate output such that we could still
verify the same behavior.

* Added new register call API

Since the the register function is part of the public API we cannot
remove/change it. However, I have marked that form of the call to be
obsolete, and instead have created a new register API which takes in
only a singular DataViewTypeAttribute. This will allow us to keep
existing functionality while still providing guidance on how the API
should actually be used. Once we increment the major version of the API
we can then remove the obsoleted version of register.

* Fixed failing test cases

While working locally, the test were not working on the CI job. this
appeared to be due to issues with path for the model, as well as not
using OnnxFact.