Fix AutoML's reference to LightGbm (#4358)
AutoML has a reference to LightGbm, however it is getting the wrong version number in the dependency. Instead of `1.4.0-preview...`, it is getting `0.16.0-preview...`. This is because the casing in the reference is incorrect.

Since the casing is incorrect, when evaluating the LightGbm project as a reference, its name no longer matches the casing in the `StableProjects` list, so it no longer thinks it is stable, and gets the unstable version.

The fix is to fix the casing, and use case-insensitve checks when evaluating IsStableProject.

Fix #4354