Brings CodeGen to master - Squash all changes to one commit (#4365)
All legs are passing except for one which appears to be a random failure and likely will not break the CI. This PR is blocking a few other important PRs that need to be in today hence I'm merging it. Spoke with @eerhardt who also agrees is ok to merge it with one leg still building and all greens.