Ensure BufferBlocks are completed and empty in RowShufflingTransformer. (#4479)
* Ensure BufferBlocks are completed and empty in RowShufflingTransformer.

If BufferBlock doesn't get completed and drained of its items, it will have non-completed Tasks. When debugging in VS, this will appear to be a memory leak because VS adds all running Tasks to a static Dictionary, and then removes them when the Task is complete. If the Task doesn't get completed, it won't be removed from the Dictionary - thus it looks like a leak.

Note there is no leak when the VS Debugger isn't attached because the non-completed Tasks don't get added to the static Dictionary.

Fix #4399