Initial featurizers project (#4413)
* DateTimeTransformer is done.

ToStringTransformer is done.

CatagoryImputer is done.

TimeSeriesImputer is done.

RobustScaler is done.

Adding in samples and documentation. General code cleanup. Made the RowToRowMapperTransform create a new mapper if possible for each cursor.

* only files needed for initial project checkin

* removed samples, fixed nuget

* changed dependencies to nightly ML.NET

* removed old TODO:

* changes from PR comments

* Fixes based on PR comments

* fixed rebase error

* Added line removed by rebase

Added line in Directory.Build.Props that was removed during the rebase process. The line already exists in master and shouldn't be removed.